revisaoalt

+++ taarefa.py
@@ -1,4 +1,4 @@
-valor = int(input("digite o valor do produto: "))
+valor = float(input("digite o valor do produto: "))
 quant = int(input("digite a quantidade: "))
 preço = valor * quant
 print(f"a sua compra deu: {preço}")
